make categories consistent
