+++ apps/main/messages/pt.po
@@ -1122,97 +1122,112 @@ msgctxt "Yp7v0H"
 msgid "Gray"
 msgstr "Cinza"
 
-#: src/lib/categories.ts
-msgctxt "4HKnmB"
-msgid "{category} Courses"
-msgstr "Cursos de {category}"
-
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "djJp6c"
 msgid "History"
 msgstr "História"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "dPgwrL"
 msgid "Math"
 msgstr "Matemática"
 
-#: src/lib/categories.ts
-msgctxt "EMi23I"
-msgid "Explore all {category} courses"
-msgstr "Explorar todos os cursos de {category}"
-
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "fFt3il"
 msgid "Arts"
 msgstr "Artes"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "GsBRWL"
 msgid "Languages"
 msgstr "Idiomas"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "hlmkcL"
 msgid "Health"
 msgstr "Saúde"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "hwL7a0"
 msgid "Engineering"
 msgstr "Engenharia"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "I1ZCPK"
 msgid "Technology"
 msgstr "Tecnologia"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "KOuDOE"
 msgid "Economics"
 msgstr "Economia"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "mn6pIm"
 msgid "Culture"
 msgstr "Cultura"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "n5D/Ku"
 msgid "Communication"
 msgstr "Comunicação"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "OV/2+4"
 msgid "Society"
 msgstr "Sociedade"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "qydxOd"
 msgid "Science"
 msgstr "Ciências"
 
-#: src/lib/categories.ts
-msgctxt "Rz9oL+"
-msgid "{category} courses"
-msgstr "Cursos de {category}"
-
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "u3ovjm"
 msgid "Law"
 msgstr "Direito"
 
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "w1Fanr"
 msgid "Business"
 msgstr "Negócios"
 
-#: src/lib/categories.ts
-msgctxt "XL8IWT"
-msgid "{category} courses. The best {category} online courses, interactive lessons to learn."
-msgstr "Cursos de {category}. Os melhores cursos online de {category}, aulas interativas pra aprender."
-
-#: src/lib/categories.ts
+#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
 msgctxt "YMDBNH"
 msgid "Geography"
 msgstr "Geografia"
+
+#: src/lib/categories/category-server.ts
+msgctxt "4HKnmB"
+msgid "{category} Courses"
+msgstr "Cursos de {category}"
+
+#: src/lib/categories/category-server.ts
+msgctxt "EMi23I"
+msgid "Explore all {category} courses"
+msgstr "Explorar todos os cursos de {category}"
+
+#: src/lib/categories/category-server.ts
+msgctxt "Rz9oL+"
+msgid "{category} courses"
+msgstr "Cursos de {category}"
+
+#: src/lib/categories/category-server.ts
+msgctxt "XL8IWT"
+msgid "{category} courses. The best {category} online courses, interactive lessons to learn."
+msgstr "Cursos de {category}. Os melhores cursos online de {category}, aulas interativas pra aprender."
